Add files via upload
Conditions:
a. Use the two switches connected to PORT A bits 1 and 0 as inputs. 
b. Based on those two inputs, make the LED D2 blink ON and OFF with the color and frequency as indicated on the table below. The LED D1 will always blink with the WHITE color
c. Place a scope probe at any pin of D1 to measure the period of that signal. The precision should be at +/-0.2 msec. Since D1 is always WHITE, any pin of D1 should reflect the blinking period of the LED D2